add cap column to patient insurance

+++ database/migrations/2023_08_15_125356_create_insurance_covers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('insurance_covers', function (Blueprint $table) {
             $table->id();
             $table->string('insurance');
-            $table->integer('cap')->nullable();
+            $table->decimal('cap', 10, 2)->nullable();
             $table->unsignedBigInteger('created_by');
             $table->foreign('created_by')->references('id')->on('users');
             $table->string('status')->default('ACTIVE');
